+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 * Improve detection of Windows target architecture when downloading prebuild dependencies. ([#6135](https://github.com/realm/realm-core/issues/6135))
 * App services request ID now logged at info level when sync client connects ([#6143](https://github.com/realm/realm-core/pull/6143]))
 * Add c_api error category for resolve errors instead of reporting unknown category. ([PR #6157](https://github.com/realm/realm-core/pull/6157))
+* Add 308 as a supported redirect response from the server ([#6162](https://github.com/realm/realm-core/issues/6162))
 
 ----------------------------------------------